backend/color_engine/requirements.txt
streamlit
pillow
numpy
opencv-python-headless
rembg
